* origin/master:
  Clean up the dead deploy pods that OSDv4 apparently doesn't do for us anymore
  Fix scaling to go off of pods under deployment rather than counting pod IPs
  Fix prometheus response coming back as a sub-sub-array
  Add ENV vars so we can modify collector resource definitions on the fly
  Change SCHEDULED_SYNC_HOURS from integer to float
  Use the safe operator when transforming source to data for ConfigMap since we do not have credentials
  Update Source#digest to account for receptor-enabled sources
  update Dockerfile to use ubi:8.2-343

+++ lib/topological_inventory/orchestrator/event_manager.rb
@@ -51,7 +51,7 @@ module TopologicalInventory
           queue.push(:event_name => "Scheduled.Sync",
                      :model      => nil)
 
-          sleep((ENV["SCHEDULED_SYNC_HOURS"]&.to_i || 1).hours)
+          sleep((ENV["SCHEDULED_SYNC_HOURS"]&.to_f || 1).hours)
         end
       end
